requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
flask_openapi3==3.1.3
numpy==2.0.0
pandas==2.2.2
pydantic==2.8.2
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.29
SQLAlchemy_Utils==0.41.2
tinydb==4.8.0
cloudpickle==3.0.0
scikit-learn==1.5.0
shap==0.46.0
matplotlib==3.9.2
pytest==8.3.3